refactor: Simplified tests
