+++ packages/ast/javascript/tests/helpers/index.ts
@@ -1,2 +1,32 @@
-export * from './javascript.js';
-export * from './typescript.js';
+import { AST } from '../../src/index.js';
+
+export function transform(file: string): string {
+  const traverse = AST.traverse();
+
+  const ast = traverse(file, {
+    visitClassDeclaration(path) {
+      const { body } = path.node.body;
+
+      const nodesToAdd = [
+        AST.builders.classProperty(
+          AST.builders.identifier('styles'),
+          AST.builders.identifier('styles'),
+        ),
+      ];
+
+      body.unshift(...nodesToAdd);
+
+      return false;
+    },
+  });
+
+  return AST.print(ast);
+}
+
+export function traverse(file: string): string {
+  const traverse = AST.traverse();
+
+  const ast = traverse(file);
+
+  return AST.print(ast);
+}
