+++ packages/ast/javascript/tests/index/javascript/transform-file-is-empty.test.ts
@@ -1,11 +1,11 @@
 import { assert, test } from '@codemod-utils/tests';
 
-import { transformJavaScript } from '../../helpers/index.js';
+import { transform } from '../../helpers/index.js';
 
 test('index | javascript > transform (file is empty)', function () {
   const oldFile = '';
 
-  const newFile = transformJavaScript(oldFile);
+  const newFile = transform(oldFile);
 
   assert.strictEqual(newFile, '');
 });
