+++ packages/ast/javascript/tests/index/javascript/traverse-file-is-empty.test.ts
@@ -1,11 +1,11 @@
 import { assert, test } from '@codemod-utils/tests';
 
-import { traverseJavaScript } from '../../helpers/index.js';
+import { traverse } from '../../helpers/index.js';
 
 test('index | javascript > traverse (file is empty)', function () {
   const oldFile = '';
 
-  const newFile = traverseJavaScript(oldFile);
+  const newFile = traverse(oldFile);
 
   assert.strictEqual(newFile, '');
 });
