+++ packages/ast/javascript/tests/index/typescript/traverse-base-case.test.ts
@@ -1,6 +1,6 @@
 import { assert, normalizeFile, test } from '@codemod-utils/tests';
 
-import { traverseTypeScript } from '../../helpers/index.js';
+import { traverse } from '../../helpers/index.js';
 
 test('index | typescript > traverse (base case)', function () {
   const oldFile = normalizeFile([
@@ -12,7 +12,7 @@ test('index | typescript > traverse (base case)', function () {
     ``,
   ]);
 
-  const newFile = traverseTypeScript(oldFile);
+  const newFile = traverse(oldFile);
 
   assert.strictEqual(
     newFile,
