+++ packages/ast/javascript/tests/index/typescript/traverse-file-is-empty.test.ts
@@ -1,11 +1,11 @@
 import { assert, test } from '@codemod-utils/tests';
 
-import { traverseTypeScript } from '../../helpers/index.js';
+import { traverse } from '../../helpers/index.js';
 
 test('index | typescript > traverse (file is empty)', function () {
   const oldFile = '';
 
-  const newFile = traverseTypeScript(oldFile);
+  const newFile = traverse(oldFile);
 
   assert.strictEqual(newFile, '');
 });
